new/requirements.txt
fastapi
uvicorn
psycopg[binary]
python-dotenv
jinja2
